create prepare_datests stage
