+++ data.dvc
@@ -1,3 +1,3 @@
 outs:
-- md5: 7dafc45044d12a46008aebfb6bf28435.dir
+- md5: 54aecb494ff2ba27236a70c1761cdcae.dir
   path: data
